The first row looks quite similar to what we’ve already seen. It finds an HTML element using the id myChart. It then gets the 2d-context of that element (the canvas) and saves it to the variable ctx. You can think of this context as a 2-dimensional whiteboard that we can use for drawing.
The rest of the code consists of a call new Chart(ctx, {...}). This is where we are telling the Chart.js tool ‘Draw a new Chart ON this 2-d whiteboard (ctx) using this data and these labels’. The type: 'line' sets the type of chart as a line chart. The label defines a label for the line. The borderColor: 'rgba(75, 192, 192, 1)' defines what colour we want the line to be. It is defined using values for red, green, blue and alpha (for transparency). The backgroundColor does the same for the fill colour of the chart.